+++ openmdao/core/problem.py
@@ -29,6 +29,8 @@ from openmdao.utils.logger_utils import get_logger
 from openmdao.utils.mpi import MPI, FakeComm
 from openmdao.utils.name_maps import prom_name2abs_name
 from openmdao.vectors.default_vector import DefaultVector
+from openmdao.jacobians.dictionary_jacobian import DictionaryJacobian
+
 try:
     from openmdao.vectors.petsc_vector import PETScVector
 except ImportError:
@@ -1406,10 +1408,10 @@ class Problem(object):
                             # need a vector for clean code, so use _views_flat.
                             dinputs._views_flat[input_name][idx - start] = 1.0
 
-                print(input_name, dinputs._views_flat[input_name])
+                # print(input_name, dinputs._views_flat[input_name])
                 model._solve_linear(lin_vec_names, mode, rel_systems)
-                #print("con1.con", doutputs._views_flat['con1.con'])
-                
+                # print("con1.con", doutputs._views_flat['con1.con'])
+
                 for input_name, old_input_name, _, matmat, simul in vois:
                     dinputs, doutputs, idxs, loc_idxs, max_i, min_i, loc_size, start, end, \
                         dup, store = voi_info[input_name]
@@ -1490,7 +1492,6 @@ class Problem(object):
                         if store and ncol > 1 and len(deriv_val.shape) == 1:
                             deriv_val = np.atleast_2d(deriv_val).T
 
-                        #print(output_name, "deriv_val", deriv_val)
                         if return_format == 'flat_dict':
                             if fwd:
                                 key = (old_output_list[ocount], old_input_name)
@@ -1849,344 +1850,3 @@ def _format_error(error, tol):
     if np.isnan(error) or error < tol:
         return '{:.6e}'.format(error)
     return '{:.6e} *'.format(error)
-
-
-def _find_var_from_range(idx, ranges):
-    # TODO: use bisection
-    for start, end, name in ranges:
-        if start <= idx <= end:
-            return name, idx - start
-
-
-def _find_disjoint(prob, mode='fwd', tol=1e-10):
-    """
-    Find all sets of disjoint columns in the total jac and their corresponding rows.
-
-    Parameters
-    ----------
-    prob : Problem
-        The Problem being analyzed.
-    mode : str
-        Derivative direction.
-    tol : float
-        Tolerance on values in jacobian.  Anything smaller in magnitude will be set to 0.0.
-
-    Returns
-    -------
-    tuple
-        Tuple of dicts total_dv_offsets and total_res_offsets.
-    """
-    # TODO: fix this to work in rev mode as well
-
-    from collections import defaultdict
-    from itertools import combinations, product
-    from openmdao.jacobians.assembled_jacobian import _SimulJacobian
-    from openmdao.utils.array_utils import array_viz
-
-    prob.model.jacobian = _SimulJacobian()
-    # clear out any old simul coloring info
-    prob.driver._simul_coloring_info = None
-    prob.driver._res_jacs = {}
-
-    prob.setup(mode=mode)
-    prob.run_model()
-
-    of = list(prob.driver._objs)
-    of_lin = []
-    desvars = prob.driver._designvars
-    responses = prob.driver._responses
-
-    wrt = list(desvars)
-    for n, meta in iteritems(prob.driver._cons):
-        if 'linear' in meta and meta['linear']:
-            of_lin.append(n)
-        else:
-            of.append(n)
-        #of.append(n)
-
-    # J = prob._compute_totals(of=of, wrt=wrt)
-    # np.set_printoptions(suppress=False)#, precision=10)
-    # print("J\n", J["phase0.collocation_constraint.defects:h","phase0.indep_controls.controls:alpha"])
-    #
-    J = prob.driver._compute_totals(return_format='array', of=of, wrt=wrt)
-    J[np.abs(J) < tol] = 0.0
-    J[np.abs(J) >= tol] = 1.0
-
-    from openmdao.utils.array_utils import array_viz
-    # array_viz(J)
-
-    # find column and row ranges (inclusive) for dvs and responses respectively
-    dv_offsets = []
-    start = 0
-    end = -1
-    for name in wrt:
-        end += prob.driver._designvars[name]['size']
-        dv_offsets.append((start, end, name))
-        # print("dv range[%s] = %s" % (name, (start, end)))
-        start = end + 1
-
-    res_offsets = []
-    start = 0
-    end = -1
-    for name in of:
-        end += responses[name]['size']
-        res_offsets.append((start, end, name))
-        # print("res range[%s] = %s" % (name, (start, end)))
-        start = end + 1
-
-    total_dv_offsets = defaultdict(lambda: defaultdict(list))
-    total_res_offsets = defaultdict(lambda: defaultdict(lambda: defaultdict(lambda: ([], []))))
-
-    # loop over each desvar and find disjoint column sets for all columns of that desvar
-    for start, end, dv in dv_offsets:
-        allcols = list(range(start, end + 1))
-        if len(allcols) < 2:
-            continue
-
-        # if dv == "phase0.indep_controls.controls:alpha":
-        #     print("size:", end-start)
-        #     with open("array.out", "w") as f:
-        #         array_viz(J[:, start:end], f)
-        #     np.set_printoptions(suppress=False)#, precision=10)
-        #     print(J[505, start:end])
-        #     res, resoffset = _find_var_from_range(505, res_offsets)
-        #     print("con:", res, resoffset)
-        #     exit()
-        disjoints = defaultdict(set)
-        rows = {}
-        for c1, c2 in combinations(allcols, 2):  # loop over column pairs
-            result = J[:, c1] + J[:, c2]
-            if np.all(result <= 1.0):
-                disjoints[c1].add(c2)
-                disjoints[c2].add(c1)
-                if c1 not in rows:
-                    rows[c1] = set(np.nonzero(J[:, c1])[0])
-                if c2 not in rows:
-                    rows[c2] = set(np.nonzero(J[:, c2])[0])
-
-        full_disjoint = {}
-        seen = set()
-        allrows = {}
-        discols = sorted(disjoints.items(), key=lambda x: len(x[1]), reverse=True)
-        for col, s in discols:
-            if col in seen or col in full_disjoint:
-                continue
-            seen.add(col)
-            allrows[col] = set(rows[col])
-            full_disjoint[col] = set([col])
-            for other_col in s:
-                if other_col not in seen and not allrows[col].intersection(rows[other_col]):
-                    seen.add(other_col)
-                    full_disjoint[col].add(other_col)
-                    allrows[col].update(rows[other_col])
-
-
-        for color, cols in enumerate(full_disjoint.values()):
-            # print("\ncolor %d:" % color)
-            for c in sorted(cols):
-                dvoffset = c - start
-                total_dv_offsets[dv][color].append(dvoffset)
-                # print(dv, dvoffset, 'col', c)
-                for crow in rows[c]:
-                    res, resoffset = _find_var_from_range(crow, res_offsets)
-                    # print("   ", res, resoffset, 'row', crow)
-                    dct = total_res_offsets[res][dv][color]
-                    dct[0].append(resoffset)
-                    dct[1].append(dvoffset)
-
-    # allcols = list(range(J.shape[1]))
-    #
-    # disjoints = defaultdict(set)
-    # rows = {}
-    # for c1, c2 in combinations(allcols, 2):  # loop over column pairs
-    #     result = J[:, c1] + J[:, c2]
-    #     if np.all(result <= 1.0):
-    #         disjoints[c1].add(c2)
-    #         disjoints[c2].add(c1)
-    #         if c1 not in rows:
-    #             rows[c1] = set(np.nonzero(J[:, c1])[0])
-    #         if c2 not in rows:
-    #             rows[c2] = set(np.nonzero(J[:, c2])[0])
-    #
-    # full_disjoint = {}
-    # seen = set()
-    # allrows = {}
-    # discols = sorted(disjoints.items(), key=lambda x: len(x[1]), reverse=True)
-    # for col, s in discols:
-    #     if col in seen or col in full_disjoint:
-    #         continue
-    #     seen.add(col)
-    #     allrows[col] = set(rows[col])
-    #     full_disjoint[col] = set([col])
-    #     for other_col in s:
-    #         if other_col not in seen and not allrows[col].intersection(rows[other_col]):
-    #             seen.add(other_col)
-    #             full_disjoint[col].add(other_col)
-    #             allrows[col].update(rows[other_col])
-    #
-    # # print("")
-    #
-    # for color, cols in enumerate(full_disjoint.values()):
-    #     # print("\ncolor %d:" % color)
-    #     for c in sorted(cols):
-    #         dv, dvoffset = _find_var_from_range(c, dv_offsets)
-    #         total_dv_offsets[dv][color].append(dvoffset)
-    #         # print(dv, dvoffset, 'col', c)
-    #         for crow in rows[c]:
-    #             res, resoffset = _find_var_from_range(crow, res_offsets)
-    #             # print("   ", res, resoffset, 'row', crow)
-    #             dct = total_res_offsets[res][dv][color]
-    #             dct[0].append(resoffset)
-    #             dct[1].append(dvoffset)
-
-    prob.driver._simul_coloring_info = None
-    prob.driver._res_jacs = {}
-
-    return total_dv_offsets, total_res_offsets
-
-
-def get_simul_meta(problem, mode='fwd', stream=sys.stdout):
-    """
-    Compute simultaneous derivative colorings for the given problem.
-
-    Parameters
-    ----------
-    problem : Problem
-        The Problem being analyzed.
-    mode : str
-        Derivative direction.
-    stream : file-like or None
-        Stream where output coloring info will be written.
-
-    Returns
-    -------
-    tuple of the form (simul_colorings, simul_maps)
-        Where simul_colorings is a dict of the form {dvname1: coloring_array, ...} and
-        simul_maps is a dict of the form
-        {resp_name: {dvname: {color: (row_idxs, col_idxs), ...}, ...}, ...}
-    """
-    driver = problem.driver
-    if not driver.supports['simultaneous_derivatives']:
-        return {}, {}
-
-    dv_idxs, res_idxs = _find_disjoint(problem, mode=mode)
-    all_colors = set()
-
-    simul_colorings = {}
-    simul_maps = {}
-
-    for dv in dv_idxs:
-        # negative colors will be iterated over individually, so start by filling the coloring array
-        # with -1.  We then replace specific entries with positive colors which will be iterated
-        # over as a group.
-        coloring = np.full(driver._designvars[dv]['size'], -1)
-        #has_color = False
-        for color in dv_idxs[dv]:
-            #if len(dv_idxs[dv][color]) > 1:
-            coloring[np.array(dv_idxs[dv][color], dtype=int)] = color
-            all_colors.add(color)
-            #has_color = True
-        #if has_color:
-        simul_colorings[dv] = list(coloring)
-
-    for res in res_idxs:
-        simul_map = {}
-        for dv in res_idxs[res]:
-            simul_map[dv] = {c: v for c, v in iteritems(res_idxs[res][dv]) if c in all_colors}
-            if not simul_map[dv]:
-                del simul_map[dv]
-
-        if simul_map:
-            simul_maps[res] = simul_map
-
-    if stream is not None:
-        stream.write("\n%s\n" % ((simul_colorings, simul_maps),))
-
-    return simul_colorings, simul_maps
-
-
-def simul_coloring_summary(problem, color_info, stream=sys.stdout):
-    """
-    Print a summary of simultaneous coloring info for the given problem and coloring metadata.
-
-    Parameters
-    ----------
-    problem : Problem
-        The Problem being analyzed.
-    color_info : tuple of (simul_colorings, simul_maps)
-        Coloring metadata.
-    stream : file-like
-        Where the output will go.
-    """
-    simul_colorings, simul_maps = color_info
-
-    desvars = problem.driver._designvars
-    responses = problem.driver._responses
-
-    tot_colors = 0
-    tot_size = 0
-    if problem._mode == 'fwd':
-        for dv in desvars:
-            if dv in simul_colorings:
-                colors = set(simul_colorings[dv])
-                if -1 in colors:
-                    negs = len(np.nonzero(np.array(simul_colorings[dv]) < 0)[0])
-                    ncolors = (negs + len(colors) - 1)
-                else:
-                    ncolors = len(colors)
-            else:
-                ncolors = desvars[dv]['size']
-            print(dv, 'num colors:', ncolors)
-            tot_colors += ncolors
-            tot_size += desvars[dv]['size']
-    else:  # rev
-        raise RuntimeError("rev mode currently not supported for simultaneous derivs.")
-
-    stream.write("\nColoring Summary\n")
-    # for dv, meta in iteritems(desvars):
-    #     stream.write("DV: %s  %d\n" % (dv, meta['size']))
-    # for res, meta in iteritems(responses):
-    #     stream.write("Resp: %s  %d\n" % (res, meta['size']))
-
-    stream.write("Total colors vs. total size: %d vs %d\n" % (tot_colors, tot_size))
-
-
-def get_total_dep(problem, dv, resp):
-    """
-    Compute total subjacobian for a given design var and response.
-
-    Note: This will raise an error if any of the components involved are matrix free.
-
-    Parameters
-    ----------
-    problem : Problem
-        Problem containing the design var and response, and all dependent vars.
-    dv : str
-        Design variable name.
-    resp : str
-        Response name.
-    """
-    inconns = problem.model._conn_global_abs_in2out
-
-    # get a dict of all outputs and their connected inputs.
-    conns = defaultdict(list)
-    for tgt, src in iteritems(inconns):
-        conns[src].append(tgt)
-
-    # get full list of components in execution order
-    comps = list(system.system_iter(include_self=True, recurse=True, typ=Component))
-    needs = {}  # dict of needed inputs vs. sys name
-    for c in comps:
-        needs[c.pathname] = set(c._var_allprocs_abs_names['input']).intersection(inconns)
-
-    relevant = problem.model._relevant
-
-    for dv in problem.driver._designvars:
-        all_ins = set()
-        all_outs = set()
-        all_sys = set()
-        for res in problem.driver._responses:
-            dct, systems = relevant[dv][res]
-            all_ins.update(dct['input'])
-            all_outs.update(dct['output'])
-            all_sys.update(systems)
