+++ openmdao/drivers/pyoptsparse_driver.py
@@ -230,8 +230,6 @@ class pyOptSparseDriver(Driver):
         lcons = [key for (key, con) in iteritems(con_meta) if con['linear'] is True]
         if len(lcons) > 0:
             _lin_jacs = self._compute_totals(of=lcons, wrt=indep_list, return_format='dict')
-            print("LIN J:\n", self._compute_totals(of=lcons, wrt=indep_list, return_format='array'))
-            #print("lin jacs", _lin_jacs)
             # convert all of our linear constraint jacs to COO format. Otherwise pyoptsparse will
             # do it for us and we'll end up with a fully dense COO matrix and very slow evaluation
             # of linear constraints!
@@ -246,7 +244,6 @@ class pyOptSparseDriver(Driver):
                         if mat.row.size > 0:
                             jacdct[n] = {'coo': [mat.row, mat.col, mat.data], 'shape': mat.shape}
                             linjac_size += mat.data.size
-            print("LIN JAC SIZE:", linjac_size)
 
         # Add all equality constraints
         self.active_tols = {}
@@ -495,10 +492,6 @@ class pyOptSparseDriver(Driver):
                 sens_dict = self._compute_totals(of=self._quantities,
                                                  wrt=self._indep_list,
                                                  return_format='dict')
-
-                #print(sens_dict['con1.con']['indep.x'],"\n")
-                #print('partial', prob.model._jacobian['con1.con', 'con1.x'])
-                #print("x", prob['indep.x'])
             # Let the optimizer try to handle the error
             except AnalysisError:
                 self._problem.model._clear_iprint()
