+++ openmdao/utils/om.py
@@ -7,7 +7,8 @@ import os
 import argparse
 from six import iteritems
 
-from openmdao.core.problem import Problem, get_simul_meta, simul_coloring_summary
+from openmdao.core.problem import Problem
+from openmdao.utils.coloring import get_simul_meta, simul_coloring_summary
 from openmdao.devtools.problem_viewer.problem_viewer import view_model
 from openmdao.devtools.viewconns import view_connections
 from openmdao.devtools.debug import config_summary, tree, dump_dist_idxs
